+++ project_frontend/src/pages/index.js
@@ -3,5 +3,6 @@ import Home from './main';
 import Welcome from './welcome';
 import PostRecruitment from './postRecruitment';
 import BoardDetails from './boardDetails';
+import PatchRecruitment from './patchRecruitment';
 
-export { Login, Home, Welcome, PostRecruitment, BoardDetails };
+export { Login, Home, Welcome, PostRecruitment, BoardDetails, PatchRecruitment };
